refactor: rewrite module fs to fs/promises
