+++ src/fs/constants.js
@@ -3,3 +3,5 @@ import { fileURLToPath } from "url";
 
 export const __filename = fileURLToPath(import.meta.url);
 export const __dirname = path.dirname(__filename);
+
+export const MESSAGE_ERROR = "FS operation failed";
