+++ src/fs/copy.js
@@ -1,21 +1,24 @@
 import path from "path";
-import { access, cp } from "fs";
+import { constants } from "fs";
+import { access, cp } from "fs/promises";
 
-import { __dirname } from "./constants.js";
+import { __dirname, MESSAGE_ERROR } from "./constants.js";
 
 const copy = async () => {
   const pathToTheFolder = path.join(__dirname, "/files");
   const pathToDestinationFolder = path.join(__dirname, "/files_copy");
 
-  access(pathToDestinationFolder, (err) => {
-    if (!err) throw new Error("FS operation failed");
-  });
-
-  cp(pathToTheFolder, pathToDestinationFolder, { recursive: true }, (err) => {
-    if (err) {
-      console.error(err);
+  try {
+    await access(pathToDestinationFolder, constants.F_OK);
+    throw new Error(MESSAGE_ERROR);
+  } catch (err) {
+    if (err.message === MESSAGE_ERROR) {
+      throw new Error(MESSAGE_ERROR);
+    } else {
+      await cp(pathToTheFolder, pathToDestinationFolder, { recursive: true });
+      console.log("Files were copied.");
     }
-  });
+  }
 };
 
 copy();
