+++ src/fs/list.js
@@ -1,15 +1,17 @@
 import path from "path";
-import { readdir } from "fs";
+import { readdir } from "fs/promises";
 
-import { __dirname } from "./constants.js";
+import { __dirname, MESSAGE_ERROR } from "./constants.js";
 
 const list = async () => {
   const pathToTheFolder = path.join(__dirname, "/files");
 
-  readdir(pathToTheFolder, (err, files) => {
-    if (err) throw new Error("FS operation failed");
+  try {
+    const files = await readdir(pathToTheFolder);
     console.log(files);
-  });
+  } catch (error) {
+    throw new Error(MESSAGE_ERROR);
+  }
 };
 
 await list();
