+++ src/fs/read.js
@@ -1,16 +1,18 @@
 import path from "path";
-import { readFile } from "fs";
+import { readFile } from "fs/promises";
 
-import { __dirname } from "./constants.js";
+import { __dirname, MESSAGE_ERROR } from "./constants.js";
 
 const read = async () => {
   const pathToTheFile = path.join(__dirname, "/files/fileToRead.txt");
 
-  readFile(pathToTheFile, (err, data) => {
-    if (err) throw new Error("FS operation failed");
+  try {
+    const content = await readFile(pathToTheFile, { encoding: "utf8" });
 
-    console.log(data.toString());
-  });
+    console.log(content);
+  } catch (err) {
+    if (err) throw new Error(MESSAGE_ERROR);
+  }
 };
 
 await read();
