+++ src/fs/rename.js
@@ -1,24 +1,18 @@
 import path from "path";
-import { readFile, rename as fsRename } from "fs";
+import { rename as giveNewName } from "fs/promises";
 
-import { __dirname } from "./constants.js";
+import { __dirname, MESSAGE_ERROR } from "./constants.js";
 
 const rename = async () => {
   const pathToTheWrongFile = path.join(__dirname, "/files/wrongFilename.txt");
   const pathToTheCorrectFile = path.join(__dirname, "/files/properFilename.md");
 
-  readFile(pathToTheWrongFile, (err, data) => {
-    if (err) throw new Error("FS operation failed");
-  });
-
-  readFile(pathToTheCorrectFile, (err, data) => {
-    if (!err) throw new Error("FS operation failed");
-  });
-
-  fsRename(pathToTheWrongFile, pathToTheCorrectFile, function (err) {
-    if (err) throw err;
-    console.log("File Renamed.");
-  });
+  try {
+    await giveNewName(pathToTheWrongFile, pathToTheCorrectFile);
+  } catch {
+    console.log("File was renamed.");
+    throw new Error(MESSAGE_ERROR);
+  }
 };
 
 await rename();
